+++ CHANGELOG.md
@@ -13,6 +13,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Security -->
 
+## [1.3.1] - 2020-05-20
+
+### Added
+
+- `ign` command
+  - Smoothing step.
+
+### Changed
+
+- `ign` command
+  - Support for regions of all shapes.
+  - Faster calculations (request parallelization).
+
+
 ## [1.3.0] - 2020-05-13
 
 ### Added